migrate: docs_examples, moon_new, output_format, target_backend, test_error_report, moon_version
