+++ crates/moon/tests/test_cases/docs_examples/avl_tree/lib/avl.mbt
