+++ crates/moon/tests/test_cases/docs_examples/avl_tree/main/main.mbt
