+++ crates/moon/tests/test_cases/docs_examples/docstring_demo/main/main.mbt
