+++ crates/moon/tests/test_cases/docs_examples/multidimensional_arrays/lib/arrays.mbt
