+++ crates/moon/tests/test_cases/docs_examples/multidimensional_arrays/main/moon.pkg.json
