+++ crates/moon/tests/test_cases/docs_examples/palindrome_string/lib/str_iter.mbt
