+++ crates/moon/tests/test_cases/docs_examples/palindrome_string/main/moon.pkg.json
