+++ crates/moon/tests/test_cases/docs_examples/palindrome_string/moon.mod.json
