+++ crates/moon/tests/test_cases/docs_examples/unicode_demo/main/moon.pkg.json
