+++ crates/moon/tests/test_cases/mod.rs
@@ -31,6 +31,7 @@ use walkdir::WalkDir;
 
 mod design;
 mod diamond_pkg;
+mod docs_examples;
 mod extra_flags;
 mod fancy_import;
 mod hello;
@@ -38,19 +39,12 @@ mod moon_bench;
 mod moon_commands;
 mod moon_new;
 mod moon_test;
+mod moon_version;
+mod output_format;
 mod simple_pkg;
+mod target_backend;
+mod test_error_report;
 mod test_filter;
-
-#[test]
-fn test_moon_run_main() {
-    let dir = TestDir::new("moon_new.in");
-    check(
-        get_stdout(&dir, ["run", "main"]),
-        expect![[r#"
-            Hello, world!
-        "#]],
-    );
-}
 
 #[test]
 fn test_moon_help() {
@@ -146,301 +140,6 @@ fn test_bench4() {
     assert!(t.contains(r#""name":"work.run""#));
     assert!(t.contains(r#""name":"run""#));
     assert!(t.contains(r#""name":"main""#));
-}
-
-#[test]
-fn test_moon_version() {
-    let dir = TestDir::new_empty();
-    let output = get_stdout(&dir, ["version"]);
-    let expected_moon_version = format!("moon {}", get_cargo_pkg_version());
-    assert!(output.contains(&expected_moon_version));
-
-    let output = get_stdout(&dir, ["version", "--all"]);
-    assert!(output.contains(&expected_moon_version));
-    assert!(output.contains("moonc"));
-    assert!(output.contains("moonrun"));
-
-    let output = get_stdout(&dir, ["version", "--all", "--no-path"]);
-    assert!(output.contains(&expected_moon_version));
-    assert!(output.contains("moonc"));
-    assert!(output.contains("moonrun"));
-}
-
-#[test]
-fn test_moon_version_json() -> anyhow::Result<()> {
-    let dir = TestDir::new_empty();
-
-    let output = get_stdout(&dir, ["version", "--json"]);
-    let items: moonutil::common::VersionItems = serde_json_lenient::from_str(&output)?;
-    assert_eq!(items.items.len(), 1);
-    assert_eq!(items.items[0].name, "moon");
-    assert!(items.items[0].version.contains(&get_cargo_pkg_version()));
-    assert!(items.items[0].path.is_some());
-
-    let output = get_stdout(&dir, ["version", "--all", "--json"]);
-    let items: moonutil::common::VersionItems = serde_json_lenient::from_str(&output)?;
-    assert_eq!(items.items.len(), 3);
-    assert_eq!(items.items[0].name, "moon");
-    assert!(items.items[0].version.contains(&get_cargo_pkg_version()));
-    assert_eq!(items.items[1].name, "moonc");
-
-    let output = get_stdout(&dir, ["version", "--all", "--json", "--no-path"]);
-    let items: moonutil::common::VersionItems = serde_json_lenient::from_str(&output)?;
-    assert!(items.items[0].path.is_none());
-
-    Ok(())
-}
-
-#[test]
-fn test_output_format() {
-    let dir = TestDir::new("output-format.in");
-
-    snapbox::cmd::Command::new(moon_bin())
-        .current_dir(&dir)
-        .args(["clean"])
-        .assert()
-        .success();
-
-    snapbox::cmd::Command::new(moon_bin())
-        .current_dir(&dir)
-        .args(["build", "-q"])
-        .assert()
-        .success();
-    assert!(dir
-        .join(format!(
-            "target/{}/release/build/main/main.wasm",
-            TargetBackend::default().to_backend_ext()
-        ))
-        .exists());
-    assert!(!dir
-        .join(format!(
-            "target/{}/release/build/main/main.wat",
-            TargetBackend::default().to_backend_ext()
-        ))
-        .exists());
-
-    snapbox::cmd::Command::new(moon_bin())
-        .current_dir(&dir)
-        .args(["clean"])
-        .assert()
-        .success();
-
-    snapbox::cmd::Command::new(moon_bin())
-        .current_dir(&dir)
-        .args(["build", "-q", "--output-wat"])
-        .assert()
-        .success();
-    assert!(dir
-        .join(format!(
-            "target/{}/release/build/main/main.wat",
-            TargetBackend::default().to_backend_ext()
-        ))
-        .exists());
-    assert!(!dir
-        .join(format!(
-            "target/{}/release/build/main/main.wasm",
-            TargetBackend::default().to_backend_ext()
-        ))
-        .exists());
-
-    snapbox::cmd::Command::new(moon_bin())
-        .current_dir(&dir)
-        .args(["clean"])
-        .assert()
-        .success();
-
-    snapbox::cmd::Command::new(moon_bin())
-        .current_dir(&dir)
-        .args(["run", "main"])
-        .assert()
-        .success();
-    assert!(!dir
-        .join(format!(
-            "target/{}/release/build/main/main.wat",
-            TargetBackend::default().to_backend_ext()
-        ))
-        .exists());
-    assert!(dir
-        .join(format!(
-            "target/{}/release/build/main/main.wasm",
-            TargetBackend::default().to_backend_ext()
-        ))
-        .exists());
-
-    snapbox::cmd::Command::new(moon_bin())
-        .current_dir(&dir)
-        .args(["clean"])
-        .assert()
-        .success();
-
-    snapbox::cmd::Command::new(moon_bin())
-        .current_dir(&dir)
-        .args(["run", "main", "--output-wat"])
-        .assert()
-        .failure();
-
-    snapbox::cmd::Command::new(moon_bin())
-        .current_dir(&dir)
-        .args(["clean"])
-        .assert()
-        .success();
-}
-
-#[test]
-fn test_target_backend() {
-    let dir = TestDir::new("target-backend.in");
-    check(
-        get_stdout(&dir, ["build", "--dry-run", "--nostd"]),
-        expect![[r#"
-            moonc build-package ./lib/hello.mbt -o ./target/wasm-gc/release/build/lib/lib.core -pkg hello/lib -pkg-sources hello/lib:./lib -target wasm-gc
-            moonc build-package ./main/main.mbt -o ./target/wasm-gc/release/build/main/main.core -pkg hello/main -is-main -i ./target/wasm-gc/release/build/lib/lib.mi:lib -pkg-sources hello/main:./main -target wasm-gc
-            moonc link-core ./target/wasm-gc/release/build/lib/lib.core ./target/wasm-gc/release/build/main/main.core -main hello/main -o ./target/wasm-gc/release/build/main/main.wasm -pkg-config-path ./main/moon.pkg.json -pkg-sources hello/lib:./lib -pkg-sources hello/main:./main -target wasm-gc
-        "#]],
-    );
-    check(
-        get_stdout(
-            &dir,
-            ["build", "--dry-run", "--target", "wasm-gc", "--nostd"],
-        ),
-        expect![[r#"
-            moonc build-package ./lib/hello.mbt -o ./target/wasm-gc/release/build/lib/lib.core -pkg hello/lib -pkg-sources hello/lib:./lib -target wasm-gc
-            moonc build-package ./main/main.mbt -o ./target/wasm-gc/release/build/main/main.core -pkg hello/main -is-main -i ./target/wasm-gc/release/build/lib/lib.mi:lib -pkg-sources hello/main:./main -target wasm-gc
-            moonc link-core ./target/wasm-gc/release/build/lib/lib.core ./target/wasm-gc/release/build/main/main.core -main hello/main -o ./target/wasm-gc/release/build/main/main.wasm -pkg-config-path ./main/moon.pkg.json -pkg-sources hello/lib:./lib -pkg-sources hello/main:./main -target wasm-gc
-        "#]],
-    );
-    check(
-        get_stdout(&dir, ["build", "--dry-run", "--target", "js", "--nostd"]),
-        expect![[r#"
-            moonc build-package ./lib/hello.mbt -o ./target/js/release/build/lib/lib.core -pkg hello/lib -pkg-sources hello/lib:./lib -target js
-            moonc build-package ./main/main.mbt -o ./target/js/release/build/main/main.core -pkg hello/main -is-main -i ./target/js/release/build/lib/lib.mi:lib -pkg-sources hello/main:./main -target js
-            moonc link-core ./target/js/release/build/lib/lib.core ./target/js/release/build/main/main.core -main hello/main -o ./target/js/release/build/main/main.js -pkg-config-path ./main/moon.pkg.json -pkg-sources hello/lib:./lib -pkg-sources hello/main:./main -target js
-        "#]],
-    );
-    check(
-        get_stdout(&dir, ["build", "--dry-run", "--nostd"]),
-        expect![[r#"
-            moonc build-package ./lib/hello.mbt -o ./target/wasm-gc/release/build/lib/lib.core -pkg hello/lib -pkg-sources hello/lib:./lib -target wasm-gc
-            moonc build-package ./main/main.mbt -o ./target/wasm-gc/release/build/main/main.core -pkg hello/main -is-main -i ./target/wasm-gc/release/build/lib/lib.mi:lib -pkg-sources hello/main:./main -target wasm-gc
-            moonc link-core ./target/wasm-gc/release/build/lib/lib.core ./target/wasm-gc/release/build/main/main.core -main hello/main -o ./target/wasm-gc/release/build/main/main.wasm -pkg-config-path ./main/moon.pkg.json -pkg-sources hello/lib:./lib -pkg-sources hello/main:./main -target wasm-gc
-        "#]],
-    );
-    check(
-        get_stdout(&dir, ["run", "main", "--dry-run", "--nostd"]),
-        expect![[r#"
-            moonc build-package ./lib/hello.mbt -o ./target/wasm-gc/release/build/lib/lib.core -pkg hello/lib -pkg-sources hello/lib:./lib -target wasm-gc
-            moonc build-package ./main/main.mbt -o ./target/wasm-gc/release/build/main/main.core -pkg hello/main -is-main -i ./target/wasm-gc/release/build/lib/lib.mi:lib -pkg-sources hello/main:./main -target wasm-gc
-            moonc link-core ./target/wasm-gc/release/build/lib/lib.core ./target/wasm-gc/release/build/main/main.core -main hello/main -o ./target/wasm-gc/release/build/main/main.wasm -pkg-config-path ./main/moon.pkg.json -pkg-sources hello/lib:./lib -pkg-sources hello/main:./main -target wasm-gc
-            moonrun ./target/wasm-gc/release/build/main/main.wasm
-        "#]],
-    );
-    check(
-        get_stdout(
-            &dir,
-            ["run", "main", "--dry-run", "--target", "wasm-gc", "--nostd"],
-        ),
-        expect![[r#"
-            moonc build-package ./lib/hello.mbt -o ./target/wasm-gc/release/build/lib/lib.core -pkg hello/lib -pkg-sources hello/lib:./lib -target wasm-gc
-            moonc build-package ./main/main.mbt -o ./target/wasm-gc/release/build/main/main.core -pkg hello/main -is-main -i ./target/wasm-gc/release/build/lib/lib.mi:lib -pkg-sources hello/main:./main -target wasm-gc
-            moonc link-core ./target/wasm-gc/release/build/lib/lib.core ./target/wasm-gc/release/build/main/main.core -main hello/main -o ./target/wasm-gc/release/build/main/main.wasm -pkg-config-path ./main/moon.pkg.json -pkg-sources hello/lib:./lib -pkg-sources hello/main:./main -target wasm-gc
-            moonrun ./target/wasm-gc/release/build/main/main.wasm
-        "#]],
-    );
-    check(
-        get_stdout(
-            &dir,
-            ["run", "main", "--dry-run", "--target", "js", "--nostd"],
-        ),
-        expect![[r#"
-            moonc build-package ./lib/hello.mbt -o ./target/js/release/build/lib/lib.core -pkg hello/lib -pkg-sources hello/lib:./lib -target js
-            moonc build-package ./main/main.mbt -o ./target/js/release/build/main/main.core -pkg hello/main -is-main -i ./target/js/release/build/lib/lib.mi:lib -pkg-sources hello/main:./main -target js
-            moonc link-core ./target/js/release/build/lib/lib.core ./target/js/release/build/main/main.core -main hello/main -o ./target/js/release/build/main/main.js -pkg-config-path ./main/moon.pkg.json -pkg-sources hello/lib:./lib -pkg-sources hello/main:./main -target js
-            node ./target/js/release/build/main/main.js
-        "#]],
-    );
-}
-
-#[test]
-fn test_test_error_report() {
-    let dir = TestDir::new("test_error_report.in");
-    snapbox::cmd::Command::new(moon_bin())
-        .current_dir(&dir)
-        .args(["test"])
-        .assert()
-        .failure();
-}
-
-#[test]
-fn test_moonbit_docs_example() {
-    let dir = TestDir::new("unicode_demo.in");
-    check(
-        get_stdout(&dir, ["run", "main"]),
-        expect![[r#"
-            3
-        "#]],
-    );
-
-    let dir = TestDir::new("palindrome_string.in");
-    check(
-        get_stdout(&dir, ["run", "main"]),
-        expect![[r#"
-        aba
-    "#]],
-    );
-
-    let dir = TestDir::new("avl_tree.in");
-    check(
-        get_stdout(&dir, ["run", "main"]),
-        expect![[r#"
-            height of the tree: 6
-                    0
-                  1
-                    2
-                3
-                    4
-                  5
-                    6
-              7
-                    8
-                  9
-                    10
-                11
-                    12
-                  13
-                    14
-            15
-                    16
-                  17
-                    18
-                19
-                    20
-                  21
-                    22
-              23
-                  24
-                25
-                    26
-                  27
-                    28
-                      29
-            success
-        "#]],
-    );
-
-    let dir = TestDir::new("docstring-demo.in");
-    check(
-        get_stdout(&dir, ["run", "main"]),
-        expect![[r#"
-            Hello, world!
-        "#]],
-    );
-
-    let dir = TestDir::new("multidimensional_arrays.in");
-    check(
-        get_stdout(&dir, ["run", "main"]),
-        expect![[r#"
-        11
-    "#]],
-    );
 }
 
 #[test]
