+++ crates/moon/tests/test_cases/output_format/lib/hello.mbt
