+++ crates/moon/tests/test_cases/output_format/main/main.mbt
