+++ crates/moon/tests/test_cases/output_format/moon.mod.json
