+++ crates/moon/tests/test_cases/target_backend/lib/hello.mbt
