+++ crates/moon/tests/test_cases/test_error_report/lib/hello.mbt
