+++ crates/moon/tests/test_cases/test_error_report/main/moon.pkg.json
